+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## Next Release
 
+## Amarok 0.2.1-rc.19
+
+- [sdk] Allow `getTransfers` query to filter by xcall caller.
+
+## Amarok 0.2.1-rc.18
+
+- [contracts] NPM package name update (@connext/contracts -> @connext/smart-contracts)
+
+## Amarok 0.2.1-rc.17
+
+- [contracts] NPM package name update (@connext/nxtp-contracts -> @connext/contracts)
+- [sdk] NPM package name update (@connext/nxtp-sdk -> @connext/sdk)
+
 ## Amarok 0.2.1-rc.16
 
 - [contracts] Added Unwrapper contract to deliver native tokens on destination